moved bindings to textcolor and style from Tab.Properties to TabView.Properties, and implemented some handler logic for android
