+++ src/library/DIPS.Mobile.UI/Components/Tabs/Android/TabHandler.cs
@@ -18,27 +18,47 @@ public class TabHandler : ViewHandler<Tab, Google.Android.Material.Tabs.TabItem>
 
     public static readonly IPropertyMapper<Tab, TabHandler> PropertyMapper = new PropertyMapper<Tab, TabHandler>(ViewMapper)
     {
-        [nameof(TabItem.Text)] = MapTitle,
+        [nameof(Tab.Title)] = MapTitle,
+        [nameof(Tab.IsSelected)] = MapIsSelected,
     };
-    
     
     protected override Google.Android.Material.Tabs.TabItem CreatePlatformView() => new(Context);
 
     protected void ConnectHandler(Google.Android.Material.Tabs.TabItem platformView)
     {
-        //base.ConnectHandler(platformView);
+        base.ConnectHandler(platformView);
         platformView.SetPadding((int)Sizes.GetSize(SizeName.content_margin_small).ToMauiPixel(), (int)Sizes.GetSize(SizeName.content_margin_xsmall).ToMauiPixel(), (int)Sizes.GetSize(SizeName.content_margin_small).ToMauiPixel(), (int)Sizes.GetSize(SizeName.content_margin_xsmall).ToMauiPixel());
-        
+        platformView.Click += OnTabTapped;
+
         var fontManager = MauiContext?.Services.GetRequiredService<IFontManager>();
     }
 
     private static void MapTitle(TabHandler handler, Tab tab)
     {
-        handler.PlatformView.Text = new Java.Lang.String(tab.Title);    
+        var counterText = tab.Counter.Length > 0 ? $" ({tab.Counter})" : string.Empty;
+        var combinedText = $"{tab.Title}{counterText}";
+
+        handler.PlatformView.Text = new Java.Lang.String(combinedText);    
+    }
+    
+    private static void MapIsSelected(TabHandler handler, Tab tab)
+    {
+        handler.PlatformView.Selected = tab.IsSelected;
+    }
+    
+    private void OnTabTapped(object? sender, EventArgs e)
+    {
+        OnTabTapped();
+    }
+    
+    internal void OnTabTapped()
+    {
+        VirtualView.SendTapped();
     }
     
     protected void DisconnectHandler(Google.Android.Material.Tabs.TabItem platformView)
     {
-        //base.DisconnectHandler(platformView);
+        base.DisconnectHandler(platformView);
+        platformView.Click -= OnTabTapped;
     }
 }
