+++ src/library/DIPS.Mobile.UI/Components/Tabs/Tab.cs
@@ -10,8 +10,5 @@ public partial class Tab : ContentView
     
     private void OnIsSelectedChanged()
     {
-#if __IOS__
-        SetTextStyle(IsSelected);
-#endif
     }
 }
